LRAC-11874 - Moving WorkspaceConnection file to default folder and Insert confirmation modal
